Shfit search functionality into its own component
